Update to be compatible with NECN
Added calculation of biomassMax for each site within the Browse extension. Previously, Browse got that information from Biomass Succession. Now, species max biomass must be provided in the input file for the Browse extension, in the same table with forage preferences, etc. The extension is now compatible with Biomass Succession and NECN Succession. The extension runs on a full landscape, but browser population dynamics and biomass consumption have not been tested whether they match original behavior of the extension.