Fixed a few issues
Fixed a problem with looping through zones when assigning defined populations.

Fixed an issue with dynamic population creating negative populations. It was also introduced when we changed the way that the initial population was input. The problem could also be caused by dividing zero if K is zero; I added a check for this and an warning if this happens.