Fixed a major bug in PartialDisturbance; other updates
PartialDisturbance was reducing cohort biomass once for every species, instead of once per site. E.g., browsed biomass was being removed 14 times because 14 species were in the species list. I'm not sure if this error existed for previous projects, but it could have a potentially huge effect on the amount of biomass removed by browsing.

I also got defoliation working, but decided against using it. Cohorts are already being reduced by browsing and the succession extension is handling what happens to that biomass. Using defoliation on top of that would double-count the foraging. It also wouldn't guarantee that the biomass removed is accurately transferred to the succession extension, because it would be limited by how much leaf biomass is available, and the forage calculation has no such limit. This would let browsers eat for free, essentially.